Work on the data structure for the store, some nomenclature changes
