+++ store/logstore.go
@@ -11,23 +11,23 @@ import (
 )
 
 // We need to include streamer on this
-type LogStore struct {
+type Logstore struct {
 	base  string
 	root  *ramstore.Dir
 	mains map[string]fs.File
 }
 
-func NewLogStore(base string, debug bool) *LogStore {
+func NewLogstore(base string, debug bool) *Logstore {
 	os.MkdirAll(base, 0777)
-	return &LogStore{
+	return &Logstore{
 		base:  base,
-		root:  ramstore.NewRoot(debug),
+		root:  ramstore.RootDir(debug),
 		mains: make(map[string]fs.File),
 	}
 
 }
 
-func (ls *LogStore) List() []string {
+func (ls *Logstore) List() []string {
 	list := ls.root.List()
 
 	for _, main := range ls.mains {
@@ -38,11 +38,11 @@ func (ls *LogStore) List() []string {
 	return list
 }
 
-func (ls *LogStore) Root(name string) (File, error) {
+func (ls *Logstore) Root(name string) (File, error) {
 	return ls.root.Root(name)
 }
 
-func (ls *LogStore) Open(name string) (File, error) {
+func (ls *Logstore) Open(name string) (File, error) {
 	// Check if our path ends with "/main"
 	if path.Base(name) == "main" || path.Base(name) == "feed" {
 		os.MkdirAll(path.Join(ls.base, path.Dir(name)), 0777)
@@ -52,7 +52,7 @@ func (ls *LogStore) Open(name string) (File, error) {
 	return ls.root.Open(name)
 }
 
-func (ls *LogStore) Delete(name string) error {
+func (ls *Logstore) Delete(name string) error {
 	if path.Base(name) == "main" || path.Base(name) == "feed" {
 		f, ok := ls.mains[name]
 		if !ok {
@@ -68,10 +68,10 @@ func (ls *LogStore) Delete(name string) error {
 	return ls.root.Delete(name)
 }
 
-func (ls *LogStore) Type() string {
+func (ls *Logstore) Type() string {
 	return "log"
 }
 
-func (ls *LogStore) Mkdir(name string) error {
+func (ls *Logstore) Mkdir(name string) error {
 	return ls.root.Mkdir(name)
 }
