+++ store/ramstore.go
@@ -14,36 +14,41 @@ const (
 )
 
 // RamStore is a type that implements Filer as an in-memory data store
-type RamStore struct {
-	root *ramstore.Dir
+type Ramstore struct {
+	Base *ramstore.Dir
 }
 
-func NewRamStore(debug bool) *RamStore {
-	return &RamStore{
-		root: ramstore.NewRoot(debug),
+func NewRamstore(debug bool) *Ramstore {
+	return &Ramstore{
+		Base: ramstore.RootDir(debug),
 	}
 }
 
-func (rs *RamStore) List() []string {
-	return rs.root.List()
+// Returns a dir or file, or any errors encountered
+func (rs *Ramstore) Walk(name string) (any, error) {
+	return rs.Base.Walk(name)
 }
 
-func (rs *RamStore) Root(name string) (File, error) {
-	return rs.root.Root(name)
+func (rs *Ramstore) List() []string {
+	return rs.Base.List()
 }
 
-func (rs *RamStore) Open(name string) (File, error) {
-	return rs.root.Open(name)
+func (rs *Ramstore) Root(name string) (File, error) {
+	return rs.Base.Root(name)
 }
 
-func (rs *RamStore) Delete(path string) error {
-	return rs.root.Delete(path)
+func (rs *Ramstore) Open(name string) (File, error) {
+	return rs.Base.Open(name)
 }
 
-func (rs *RamStore) Type() string {
+func (rs *Ramstore) Delete(path string) error {
+	return rs.Base.Delete(path)
+}
+
+func (rs *Ramstore) Type() string {
 	return "ram"
 }
 
-func (rs *RamStore) Mkdir(dir string) error {
-	return rs.root.Mkdir(dir)
+func (rs *Ramstore) Mkdir(dir string) error {
+	return rs.Base.Mkdir(dir)
 }
